backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
huggingface_hub==0.34.4c;
python-multipart==0.0.6
pydantic==2.5.0